Finished standalone batch/run tests for model.
* Factored out simple model into separate source/exe files.
* Put together simple standalone tests for run/batch model commands.
* Fixed a couple of bugs in text_buffer(!).
* Fixed some oversights in model.
