+++ model/tests/simple.c
@@ -1,0 +1,14 @@
+// Copyright (c) 2012-2016, Jeffrey N. Johnson
+// All rights reserved.
+// 
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+#include "simple_model.h"
+
+int main(int argc, char* argv[])
+{
+  return model_main("simple", simple_new, argc, argv);
+}
+
