+++ model/tests/simple_model.h
@@ -1,0 +1,12 @@
+// Copyright (c) 2012-2016, Jeffrey N. Johnson
+// All rights reserved.
+// 
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+#include "model/model.h"
+
+// This creates a simple model for testing.
+model_t* simple_new(void);
+
